--- .babelrc
@@ -1,70 +1,0 @@
-{
-  "env": {
-    "dev": {
-      "presets": [
-        [
-          "@babel/preset-env",
-
-          {
-            "modules": false,
-            "targets": {
-              "firefox": "103",
-              "chrome": "104"
-            }
-          }
-        ],
-        [
-          "@babel/preset-react",
-          {
-            "runtime": "automatic"
-          }
-        ],
-        "@babel/preset-typescript"
-      ]
-    },
-
-    "cjs": {
-      "presets": [
-        "@babel/preset-env",
-        [
-          "@babel/preset-react",
-          {
-            "runtime": "automatic"
-          }
-        ],
-        "@babel/preset-typescript"
-      ]
-    },
-
-    "es": {
-      "presets": [
-        [
-          "@babel/preset-env",
-          {
-            "modules": false
-          }
-        ],
-        [
-          "@babel/preset-react",
-          {
-            "runtime": "automatic"
-          }
-        ],
-        "@babel/preset-typescript"
-      ]
-    },
-
-    "test": {
-      "presets": [
-        "@babel/preset-env",
-        "@babel/preset-react",
-        "@babel/preset-typescript"
-      ]
-    }
-  },
-
-  "plugins": [
-    "@babel/plugin-transform-runtime",
-    "@babel/plugin-proposal-class-properties"
-  ]
-}
