+++ package.json
@@ -2,24 +2,30 @@
   "name": "@vtaits/react-color-picker",
   "version": "0.1.3",
   "description": "React Color Picker",
-  "main": "lib/index.js",
-  "module": "es/index.js",
   "files": [
-    "/dist",
-    "/es",
-    "/lib"
+    "dist"
   ],
+  "main": "./dist/index.js",
+  "module": "./dist/esm/index.js",
+  "exports": {
+    ".": {
+      "import": "./dist/esm/index.js",
+      "require": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
+  },
+  "typings": "dist/index.d.ts",
   "scripts": {
+    "storybook": "storybook dev -p 6006",
     "start": "cross-env BABEL_ENV=dev webpack-dev-server",
-    "build": "yarn build:cjs && yarn build:es && yarn build:css",
-    "build:cjs": "cross-env BABEL_ENV=cjs babel src --out-dir lib --ignore \"src/**/__tests__\"",
-    "build:es": "cross-env BABEL_ENV=es babel src --out-dir es --ignore \"src/**/__tests__\"",
+    "build": "yarn build:js && yarn build:css",
+    "build:js": "tsup src/index.ts --sourcemap --format esm,cjs --dts --legacy-output",
     "build:css": "postcss style/**/*.css --dir dist",
-    "clean": "rm -rf lib es dist",
-    "lib": "yarn clean && yarn babel",
-    "babel": "babel --out-dir lib src",
+    "clean": "rm -rf dist",
     "prepare": "yarn clean && yarn build",
-    "lint": "BABEL_ENV=es eslint src --ext .js,.jsx",
+    "lint": "biome check src",
+    "lint:fix": "biome check src --apply-unsafe",
+    "format": "biome format src --write",
     "test": "yarn lint"
   },
   "keywords": [
@@ -30,46 +36,38 @@
     "color-picker"
   ],
   "dependencies": {
-    "@babel/runtime": "^7.18.9",
-    "@types/react": "^18.0.17",
-    "classnames": "^2.3.1",
+    "@types/react": "^18.2.34",
+    "classnames": "^2.3.2",
     "drag-helper": "^1.3.6",
-    "prop-types": "^15.8.1",
     "region": "^2.1.2",
-    "tinycolor2": "^1.4.2"
+    "tinycolor2": "^1.6.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.18.10",
-    "@babel/core": "^7.18.10",
-    "@babel/eslint-parser": "^7.18.9",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-transform-runtime": "^7.18.10",
-    "@babel/preset-env": "^7.18.10",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.18.6",
-    "@typescript-eslint/eslint-plugin": "^5.33.1",
-    "@typescript-eslint/parser": "^5.33.1",
-    "autoprefixer": "^10.4.8",
-    "babel-loader": "^8.2.5",
-    "cross-env": "^7.0.3",
-    "css-loader": "^6.7.1",
-    "eslint": "^8.22.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-import-resolver-typescript": "^3.4.1",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
-    "eslint-plugin-react": "^7.30.1",
-    "postcss": "^8.4.16",
-    "postcss-cli": "^10.0.0",
+    "@biomejs/biome": "^1.3.3",
+    "@storybook/addon-essentials": "^7.5.2",
+    "@storybook/addon-interactions": "^7.5.2",
+    "@storybook/addon-links": "^7.5.2",
+    "@storybook/blocks": "^7.5.2",
+    "@storybook/builder-vite": "^7.5.2",
+    "@storybook/manager-api": "^7.5.2",
+    "@storybook/react": "^7.5.2",
+    "@storybook/react-vite": "^7.5.2",
+    "@storybook/testing-library": "^0.2.2",
+    "@types/tinycolor2": "^1.4.5",
+    "postcss": "^8.4.31",
+    "postcss-cli": "^10.1.0",
     "precss": "^4.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "style-loader": "^3.3.1",
-    "typescript": "^4.7.4",
-    "webpack": "^5.74.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.10.0"
+    "storybook": "^7.5.2",
+    "tsup": "^7.2.0",
+    "typescript": "^5.2.2",
+    "vite": "^4.5.0"
+  },
+  "resolutions": {
+    "string-width": "^4",
+    "jackspeak": "^2",
+    "strip-ansi": "^6"
   },
   "peerDependencies": {
     "react": "^16.3.0 || ^17.0.0 || ^18.0.0",
