+++ postcss.config.js
@@ -1,9 +1,7 @@
-const autoprefixer = require('autoprefixer');
 const precss = require('precss');
 
 module.exports = {
   plugins: [
     precss(),
-    autoprefixer(),
   ],
 };
