+++ src/defaultColor.ts
@@ -1,1 +1,1 @@
-export const DEFAULT_COLOR = 'red';
+export const DEFAULT_COLOR = "red";
