+++ src/index.ts
@@ -1,0 +1,3 @@
+export { ColorPicker } from "./ColorPicker";
+export { HueSpectrum } from "./HueSpectrum";
+export { SaturationSpectrum } from "./SaturationSpectrum";
