+++ tsconfig.json
@@ -1,11 +1,28 @@
 {
   "compilerOptions": {
-    "lib": ["es2020", "DOM"],
-    "target": "es5",
-    "outDir": "ts",
+    "target": "ES2021",
+    "outDir": "dist/ts",
+    "jsx": "react-jsx",
+    "lib": [
+      "ESNext",
+      "DOM",
+      "DOM.Iterable"
+    ],
+    "allowJs": false,
+    "skipLibCheck": true,
     "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
     "strict": true,
-    "strictNullChecks": true,
-    "jsx": "react-jsx"
-  }
+    "forceConsistentCasingInFileNames": true,
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
+    "resolveJsonModule": true,
+    "isolatedModules": true
+  },
+  "include": [
+    "src/**/*"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
 }
